+++ clients/web3-wallet/src/main/java/com/binance/connector/client/web3_wallet/rest/api/RwaDataApi.java
@@ -1,6 +1,6 @@
 /*
  * Binance Web3 API
- * Cross-chain wallet, market, trading, and transaction APIs for the Binance Web3 API platform.
+ * Cross-chain wallet, market, trading, transaction, and B402 payment APIs for the Binance Web3 API platform.
  *
  * The version of the OpenAPI document: 1.0.0
  *
@@ -48,7 +48,7 @@ public class RwaDataApi {
 
     private static final String USER_AGENT =
             String.format(
-                    "binance-web3-wallet/4.1.0 (Java/%s; %s; %s)",
+                    "binance-web3-wallet/4.2.0 (Java/%s; %s; %s)",
                     SystemUtil.getJavaVersion(), SystemUtil.getOs(), SystemUtil.getArch());
     private static final boolean HAS_TIME_UNIT = false;